Deleted unused PrintedProduct service, Status value object, and multiple obsolete database migration files to clean up the codebase and reduce technical debt. This simplifies maintenance and focuses on current business logic.

+++ src/Orders/Application/UseCase/Command/AddProduct/AddProductCommand.php
@@ -22,7 +22,7 @@ readonly class AddProductCommand implements CommandInterface
      */
     public function __construct(
         public int $orderId,
-        public int $length,
+        public float $length,
         public ?string $filmType,
         public ?string $laminationType = null,
         public ?int $cutFileId = null,
